ADD getting rooms datas for a capsule
